+++ tests/test_extremes.py
@@ -1,21 +1,17 @@
-import os
-import warnings
-
 import numpy as np
-import pandas as pd
 import pytest
 import xarray as xr
-
-from xhydro.extreme_value_analysis.parameterestimation import fit, return_level
-
-# pytest -n0 --cov=src\xhydro --cov-report=html tests/test_extremes.py
 
 
 @pytest.mark.requires_julia
 class Testfit:
 
     def test_stationary_grv(self, data_fre):
-        p = fit(data_fre, dist="genextreme", method="ml", vars=["SeaLevel"], dim="Year")
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.fit(
+            data_fre, dist="genextreme", method="ml", variables=["SeaLevel"], dim="Year"
+        )
 
         assert isinstance(p, xr.Dataset)
         assert list(p.coords) == ["dparams"]
@@ -42,7 +38,11 @@ class Testfit:
         )
 
     def test_stationary_gumbel(self, data_fre):
-        p = fit(data_fre, dist="gumbel_r", method="ml", vars=["SeaLevel"], dim="Year")
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.fit(
+            data_fre, dist="gumbel_r", method="ml", variables=["SeaLevel"], dim="Year"
+        )
 
         assert isinstance(p, xr.Dataset)
         assert list(p.coords) == ["dparams"]
@@ -69,8 +69,14 @@ class Testfit:
         )
 
     def test_stationary_pareto(self, data_rain):
-        p = fit(
-            data_rain, dist="genpareto", method="ml", vars=["Exceedance"], dim="Date"
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.fit(
+            data_rain,
+            dist="genpareto",
+            method="ml",
+            variables=["Exceedance"],
+            dim="Date",
         )
 
         assert isinstance(p, xr.Dataset)
@@ -98,11 +104,13 @@ class Testfit:
         )
 
     def test_non_stationary_cov(self, data_fre):
-        p = fit(
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.fit(
             data_fre,
             dist="genextreme",
             method="ml",
-            vars=["SeaLevel"],
+            variables=["SeaLevel"],
             dim="Year",
             locationcov=["Year"],
         )
@@ -133,11 +141,13 @@ class Testfit:
         )
 
     def test_non_stationary_cov2(self, data_fre):
-        p = fit(
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.fit(
             data_fre,
             dist="genextreme",
             method="ml",
-            vars=["SeaLevel"],
+            variables=["SeaLevel"],
             dim="Year",
             locationcov=["Year", "SOI"],
             scalecov=["Year", "SOI"],
@@ -203,17 +213,18 @@ class Testfit:
         )
 
     def test_non_stationary_cov_pwm(self, data_fre):
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
 
         with pytest.raises(
             ValueError,
             match="Probability weighted moment parameter estimation cannot have covariates \\['Year', 'SOI', 'Year2', 'SOI2', 'Year3', 'SOI3'\\]",
         ):
 
-            fit(
+            evap.fit(
                 data_fre,
                 dist="genextreme",
                 method="pwm",
-                vars=["SeaLevel"],
+                variables=["SeaLevel"],
                 dim="Year",
                 locationcov=["Year", "SOI"],
                 scalecov=["Year2", "SOI2"],
@@ -221,17 +232,18 @@ class Testfit:
             )
 
     def test_fewer_entries_than_param_param(self, data_fre):
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
 
         with pytest.warns(
             UserWarning,
             match="The fitting data contains fewer entries than the number of parameters for the given distribution. "
             "Returned parameters are numpy.nan",
         ):
-            p = fit(
+            p = evap.fit(
                 data_fre.isel(Year=slice(1, 2)),
                 dist="genextreme",
                 method="ml",
-                vars=["SeaLevel"],
+                variables=["SeaLevel"],
                 dim="Year",
                 locationcov=["Year", "SOI"],
                 scalecov=["Year", "SOI"],
@@ -260,17 +272,18 @@ class Testfit:
         )
 
     def test_fewer_entries_than_param_rtnlev(self, data_fre):
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
 
         with pytest.warns(
             UserWarning,
             match="The fitting data contains fewer entries than the number of parameters for the given distribution. "
             "Returned parameters are numpy.nan",
         ):
-            p = return_level(
+            p = evap.return_level(
                 data_fre.isel(Year=slice(0, 5)),
                 dist="genextreme",
                 method="ml",
-                vars=["SeaLevel"],
+                variables=["SeaLevel"],
                 dim="Year",
                 locationcov=["Year", "SOI"],
                 scalecov=["Year", "SOI"],
@@ -292,12 +305,13 @@ class Testfit:
 class TestRtnlv:
 
     def test_stationary(self, data_fre):
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
 
-        p = return_level(
+        p = evap.return_level(
             data_fre,
             dist="genextreme",
             method="ml",
-            vars=["SeaLevel"],
+            variables=["SeaLevel"],
             dim="Year",
             return_period=100,
         )
@@ -321,12 +335,13 @@ class TestRtnlv:
         )
 
     def test_non_stationary_cov(self, data_fre):
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
 
-        p = return_level(
+        p = evap.return_level(
             data_fre,
             dist="genextreme",
             method="ml",
-            vars=["SeaLevel"],
+            variables=["SeaLevel"],
             dim="Year",
             locationcov=["Year", "SOI"],
             return_period=100,
@@ -358,16 +373,17 @@ class TestRtnlv:
         )
 
     def test_recover_nan(self, data_fre):
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
 
         data_fre_nan = data_fre.copy(deep=True)
         data_fre_nan["SeaLevel"].isel(Year=slice(1, 3))[:] = np.nan
         data_fre_nan["SOI"].isel(Year=slice(5, 6))[:] = np.nan
 
-        p = return_level(
+        p = evap.return_level(
             data_fre_nan,
             dist="genextreme",
             method="ml",
-            vars=["SeaLevel"],
+            variables=["SeaLevel"],
             dim="Year",
             locationcov=["Year", "SOI"],
             return_period=100,
@@ -423,7 +439,11 @@ class TestRtnlv:
 class TestGEV:
 
     def test_ml_param(self, data_fre):
-        p = fit(data_fre, dist="genextreme", method="ML", dim="Year", vars=["SeaLevel"])
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.fit(
+            data_fre, dist="genextreme", method="ML", dim="Year", variables=["SeaLevel"]
+        )
 
         assert (
             (p["SeaLevel_lower"] < p["SeaLevel"])
@@ -441,8 +461,14 @@ class TestGEV:
         )
 
     def test_pwm_param(self, data_fre):
-        p = fit(
-            data_fre, dist="genextreme", method="PWM", dim="Year", vars=["SeaLevel"]
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.fit(
+            data_fre,
+            dist="genextreme",
+            method="PWM",
+            dim="Year",
+            variables=["SeaLevel"],
         )
 
         assert (
@@ -454,8 +480,14 @@ class TestGEV:
         )
 
     def test_bayes_param(self, data_fre):
-        p = fit(
-            data_fre, dist="genextreme", method="BAYES", dim="Year", vars=["SeaLevel"]
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.fit(
+            data_fre,
+            dist="genextreme",
+            method="BAYES",
+            dim="Year",
+            variables=["SeaLevel"],
         )
 
         assert (
@@ -464,8 +496,10 @@ class TestGEV:
         ).values.all()
 
     def test_ml_rtnlv(self, data_fre):
-        p = return_level(
-            data_fre, dist="genextreme", method="ML", dim="Year", vars=["SeaLevel"]
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.return_level(
+            data_fre, dist="genextreme", method="ML", dim="Year", variables=["SeaLevel"]
         )
 
         assert (
@@ -477,8 +511,14 @@ class TestGEV:
         )
 
     def test_pwm_rtnlv(self, data_fre):
-        p = return_level(
-            data_fre, dist="genextreme", method="PWM", dim="Year", vars=["SeaLevel"]
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.return_level(
+            data_fre,
+            dist="genextreme",
+            method="PWM",
+            dim="Year",
+            variables=["SeaLevel"],
         )
 
         np.testing.assert_array_almost_equal(
@@ -490,8 +530,14 @@ class TestGEV:
         ).values.all()
 
     def test_bayes_rtnlv(self, data_fre):
-        p = return_level(
-            data_fre, dist="genextreme", method="BAYES", dim="Year", vars=["SeaLevel"]
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.return_level(
+            data_fre,
+            dist="genextreme",
+            method="BAYES",
+            dim="Year",
+            variables=["SeaLevel"],
         )
 
         assert (
@@ -504,7 +550,11 @@ class TestGEV:
 class TestGumbel:
 
     def test_ml_param(self, data_fre):
-        p = fit(data_fre, dist="gumbel_r", method="ML", dim="Year", vars=["SeaLevel"])
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.fit(
+            data_fre, dist="gumbel_r", method="ML", dim="Year", variables=["SeaLevel"]
+        )
 
         assert (
             (p["SeaLevel_lower"] < p["SeaLevel"])
@@ -522,7 +572,11 @@ class TestGumbel:
         )
 
     def test_pwm_param(self, data_fre):
-        p = fit(data_fre, dist="gumbel_r", method="PWM", dim="Year", vars=["SeaLevel"])
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.fit(
+            data_fre, dist="gumbel_r", method="PWM", dim="Year", variables=["SeaLevel"]
+        )
 
         assert (
             (p["SeaLevel_lower"] < p["SeaLevel"])
@@ -533,8 +587,14 @@ class TestGumbel:
         )
 
     def test_bayes_param(self, data_fre):
-        p = fit(
-            data_fre, dist="gumbel_r", method="BAYES", dim="Year", vars=["SeaLevel"]
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.fit(
+            data_fre,
+            dist="gumbel_r",
+            method="BAYES",
+            dim="Year",
+            variables=["SeaLevel"],
         )
 
         assert (
@@ -543,8 +603,10 @@ class TestGumbel:
         ).values.all()
 
     def test_ml_rtnlv(self, data_fre):
-        p = return_level(
-            data_fre, dist="gumbel_r", method="ML", dim="Year", vars=["SeaLevel"]
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.return_level(
+            data_fre, dist="gumbel_r", method="ML", dim="Year", variables=["SeaLevel"]
         )
 
         assert (
@@ -556,8 +618,10 @@ class TestGumbel:
         )
 
     def test_pwm_rtnlv(self, data_fre):
-        p = return_level(
-            data_fre, dist="gumbel_r", method="PWM", dim="Year", vars=["SeaLevel"]
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.return_level(
+            data_fre, dist="gumbel_r", method="PWM", dim="Year", variables=["SeaLevel"]
         )
 
         np.testing.assert_array_almost_equal(
@@ -569,8 +633,14 @@ class TestGumbel:
         ).values.all()
 
     def test_bayes_rtnlv(self, data_fre):
-        p = return_level(
-            data_fre, dist="gumbel_r", method="BAYES", dim="Year", vars=["SeaLevel"]
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.return_level(
+            data_fre,
+            dist="gumbel_r",
+            method="BAYES",
+            dim="Year",
+            variables=["SeaLevel"],
         )
 
         assert (
@@ -583,8 +653,14 @@ class TestGumbel:
 class TestPareto:
 
     def test_ml_param(self, data_rain):
-        p = fit(
-            data_rain, dist="genpareto", method="ML", dim="Date", vars=["Exceedance"]
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.fit(
+            data_rain,
+            dist="genpareto",
+            method="ML",
+            dim="Date",
+            variables=["Exceedance"],
         )
 
         assert (
@@ -603,8 +679,14 @@ class TestPareto:
         )
 
     def test_pwm_param(self, data_rain):
-        p = fit(
-            data_rain, dist="genpareto", method="PWM", dim="Date", vars=["Exceedance"]
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.fit(
+            data_rain,
+            dist="genpareto",
+            method="PWM",
+            dim="Date",
+            variables=["Exceedance"],
         )
 
         assert (
@@ -616,8 +698,14 @@ class TestPareto:
         )
 
     def test_bayes_param(self, data_rain):
-        p = fit(
-            data_rain, dist="genpareto", method="BAYES", vars=["Exceedance"], dim="Date"
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.fit(
+            data_rain,
+            dist="genpareto",
+            method="BAYES",
+            variables=["Exceedance"],
+            dim="Date",
         )
 
         assert (
@@ -626,12 +714,14 @@ class TestPareto:
         ).values.all()
 
     def test_ml_rtnlv(self, data_rain):
-        p = return_level(
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.return_level(
             data_rain,
             dist="genpareto",
             method="ML",
             dim="Date",
-            vars=["Exceedance"],
+            variables=["Exceedance"],
             threshold_pareto=30,
             nobs_pareto=17531,
             nobsperblock_pareto=365,
@@ -647,12 +737,14 @@ class TestPareto:
         )
 
     def test_pwm_rtnlv(self, data_rain):
-        p = return_level(
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.return_level(
             data_rain,
             dist="genpareto",
             method="PWM",
             dim="Date",
-            vars=["Exceedance"],
+            variables=["Exceedance"],
             threshold_pareto=30,
             nobs_pareto=17531,
             nobsperblock_pareto=365,
@@ -666,12 +758,14 @@ class TestPareto:
         ).values.all()
 
     def test_bayes_rtnlv(self, data_rain):
-        p = return_level(
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
+        p = evap.return_level(
             data_rain,
             dist="genpareto",
             method="BAYES",
             dim="Date",
-            vars=["Exceedance"],
+            variables=["Exceedance"],
             threshold_pareto=30,
             nobs_pareto=17531,
             nobsperblock_pareto=365,
@@ -686,18 +780,23 @@ class TestPareto:
 @pytest.mark.requires_julia
 class TestError:
 
+    dataset = xr.Dataset(
+        {"n": ("pos", [1, -1, 1, -1, 1, -1]), "n2": ("pos", [2, -2, 2, -2, 2, -2])},
+        coords={"pos": [0, 1, 2, 3, 4, -5]},
+    )
+
     def test_non_stationary_cov_pwm(self, data_fre):
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
 
         with pytest.raises(
             ValueError,
             match="Probability weighted moment parameter estimation cannot have covariates \\['Year', 'SOI', 'Year2', 'SOI2', 'Year3', 'SOI3'\\]",
         ):
-
-            return_level(
+            evap.return_level(
                 data_fre,
                 dist="genextreme",
                 method="pwm",
-                vars=["SeaLevel"],
+                variables=["SeaLevel"],
                 dim="Year",
                 locationcov=["Year", "SOI"],
                 scalecov=["Year2", "SOI2"],
@@ -705,38 +804,47 @@ class TestError:
             )
 
     def test_extremefit_rtnlv_error(self, data_rain):
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
 
         with pytest.raises(ValueError, match="Unrecognized distribution: XXX"):
-            return_level(
+            evap.return_level(
                 data_rain,
                 dist="XXX",
                 method="BAYES",
                 dim="Date",
-                vars=["Exceedance"],
+                variables=["Exceedance"],
                 threshold_pareto=30,
                 nobs_pareto=17531,
                 nobsperblock_pareto=365,
             )
 
     def test_extremefit_param_error(self, data_rain):
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
         with pytest.raises(ValueError, match="Unrecognized method: YYY"):
-            return_level(
+            evap.return_level(
                 data_rain,
                 dist="genpareto",
                 method="YYY",
                 dim="Date",
-                vars=["Exceedance"],
+                variables=["Exceedance"],
                 threshold_pareto=30,
                 nobs_pareto=17531,
                 nobsperblock_pareto=365,
             )
 
     def test_confinc_error(self, data_fre):
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
         with pytest.warns(
             UserWarning, match="There was an error in computing confidence interval."
         ):
-            p = fit(
-                data_fre, dist="genpareto", method="ML", vars=["SeaLevel"], dim="Year"
+            p = evap.fit(
+                data_fre,
+                dist="genpareto",
+                method="ML",
+                variables=["SeaLevel"],
+                dim="Year",
             )
 
         np.testing.assert_array_almost_equal(
@@ -745,23 +853,21 @@ class TestError:
         )
 
     def test_confinc_error_nostat(self):
-        dataset = xr.Dataset(
-            {"n": ("pos", [1, -1, 1, -1, 1, -1]), "n2": ("pos", [2, -2, 2, -2, 2, -2])},
-            coords={"pos": [0, 1, 2, 3, 4, -5]},
-        )
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
         with pytest.warns(
             UserWarning,
             match="There was an error in fitting the data to a genpareto distribution using ML. "
             "Returned parameters are numpy.nan",
         ):
 
-            p = fit(
-                dataset,
+            p = evap.fit(
+                self.dataset,
                 dist="genpareto",
                 method="ML",
                 dim="pos",
                 scalecov=["n2"],
-                vars=["n"],
+                variables=["n"],
             )
             np.testing.assert_array_equal(
                 p.to_array().values,
@@ -775,23 +881,21 @@ class TestError:
             )
 
     def test_confinc_error_nostat_rtnlv(self):
-        dataset = xr.Dataset(
-            {"n": ("pos", [1, -1, 1, -1, 1, -1]), "n2": ("pos", [2, -2, 2, -2, 2, -2])},
-            coords={"pos": [0, 1, 2, 3, 4, -5]},
-        )
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
         with pytest.warns(
             UserWarning,
             match="There was an error in fitting the data to a genpareto distribution using BAYES. "
             "Returned parameters are numpy.nan",
         ):
 
-            p = fit(
-                dataset,
+            p = evap.fit(
+                self.dataset,
                 dist="genpareto",
                 method="BAYES",
                 dim="pos",
                 scalecov=["n2"],
-                vars=["n"],
+                variables=["n"],
             )
 
             np.testing.assert_array_equal(
@@ -806,23 +910,21 @@ class TestError:
             )
 
     def test_rtnlv_error_nostat_rtnlv(self):
-        dataset = xr.Dataset(
-            {"n": ("pos", [1, -1, 1, -1, 1, -1]), "n2": ("pos", [2, -2, 2, -2, 2, -2])},
-            coords={"pos": [0, 1, 2, 3, 4, -5]},
-        )
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
         with pytest.warns(
             UserWarning,
             match="There was an error in fitting the data to a genextreme distribution using BAYES. "
             "Returned parameters are numpy.nan",
         ):
 
-            p = return_level(
-                dataset,
+            p = evap.return_level(
+                self.dataset,
                 dist="genextreme",
                 method="BAYES",
                 dim="pos",
                 scalecov=["n2"],
-                vars=["n"],
+                variables=["n"],
             )
 
             np.testing.assert_array_equal(
@@ -837,85 +939,107 @@ class TestError:
             )
 
     def test_dist_error(self, data_fre):
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
         with pytest.raises(ValueError, match="Unrecognized distribution: XXX"):
-            fit(data_fre, dist="XXX", method="ml", vars=["SeaLevel"], dim="Year")
+            evap.fit(
+                data_fre, dist="XXX", method="ml", variables=["SeaLevel"], dim="Year"
+            )
 
     def test_method_error(self, data_fre):
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
         with pytest.raises(ValueError, match="Unrecognized method: YYY"):
-            fit(
-                data_fre, dist="genextreme", method="YYY", vars=["SeaLevel"], dim="Year"
+            evap.fit(
+                data_fre,
+                dist="genextreme",
+                method="YYY",
+                variables=["SeaLevel"],
+                dim="Year",
             )
 
     def test_vars_error(self, data_fre):
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
         with pytest.raises(
             ValueError,
-            match="XXXX is not a variable in the Dataset. Dataset's variables are: \\['SOI', 'SeaLevel'\\]",
+            match="XXXX is not a variable in the Dataset. Dataset variables are: \\['SOI', 'SeaLevel'\\]",
         ):
-            fit(data_fre, dist="genextreme", vars=["XXXX"], dim="Year")
+            evap.fit(data_fre, dist="genextreme", variables=["XXXX"], dim="Year")
 
     def test_conflev_error(self, data_fre):
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
         with pytest.raises(
             ValueError,
             match="Confidence level must be strictly smaller than 1 and strictly larger than 0",
         ):
-            fit(
+            evap.fit(
                 data_fre,
                 dist="genextreme",
-                vars=["SeaLevel"],
+                variables=["SeaLevel"],
                 dim="Year",
                 confidence_level=2,
             )
 
     def test_gumbel_cov_error(self, data_fre):
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
         with pytest.raises(
             ValueError,
             match="Gumbel distribution has no shape parameter and thus cannot have shape covariates \\['Year'\\]",
         ):
-            fit(
+            evap.fit(
                 data_fre,
                 dist="gumbel_r",
                 method="ml",
-                vars=["SeaLevel"],
+                variables=["SeaLevel"],
                 dim="Year",
                 shapecov=["Year"],
             )
 
     def test_pareto_cov_error(self, data_fre):
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
         with pytest.raises(
             ValueError,
             match="Pareto distribution has no location parameter and thus cannot have location covariates \\['Year'\\]",
         ):
-            return_level(
+            evap.return_level(
                 data_fre,
                 dist="genpareto",
                 method="ml",
-                vars=["SeaLevel"],
+                variables=["SeaLevel"],
                 dim="Year",
                 locationcov=["Year"],
             )
 
     def test_rtmper_error(self, data_fre):
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
         with pytest.raises(
             ValueError,
             match="Return period has to be strictly larger than 0. Current return period value is -1",
         ):
-            return_level(
+            evap.return_level(
                 data_fre,
                 dist="genextreme",
-                vars=["SeaLevel"],
+                variables=["SeaLevel"],
                 dim="Year",
                 return_period=-1,
             )
 
     def test_para_pareto_error(self, data_rain):
+        evap = pytest.importorskip("xhydro.extreme_value_analysis.parameterestimation")
+
         with pytest.raises(
             ValueError,
             match="'threshold_pareto', 'nobs_pareto', and 'nobsperblock_pareto' must be defined when using dist 'genpareto'.",
         ):
-            return_level(
+            evap.return_level(
                 data_rain,
                 dist="genpareto",
                 method="ml",
-                vars=["Exceedance"],
+                variables=["Exceedance"],
                 dim="Date",
             )